Fix Progress page crash: intervalDays cast to int broke on real fractional values

+++ apps/mobile/lib/features/home/data/models/home_review_schedule_model.dart
@@ -27,7 +27,7 @@ class HomeReviewScheduleModel extends HomeReviewSchedule {
       scheduleId: json['scheduleId'] as String,
       skillId: json['skillId'] as String,
       dueAt: json['dueAt'] as String,
-      intervalDays: (json['intervalDays'] as num).toInt(),
+      intervalDays: (json['intervalDays'] as num).toDouble(),
       repetitionCount: (json['repetitionCount'] as num).toInt(),
       status: json['status'] as String,
       basedOnAttemptId: json['basedOnAttemptId'] as String,
